添加商品页面步骤条

+++ src/views/goodsCom/goods/Goods.vue
@@ -1,3 +1,9 @@
+/*
+ * @Author: wangxin.leo
+ * @Date: 2020-05-22 14:38:51
+ * @Last Modified by: wangxin.leo
+ * @Last Modified time: 2020-05-22 15:26:41
+ */
 <template>
   <div class="goods">
     <!--    面包屑导航-->
@@ -6,13 +12,15 @@
       <el-breadcrumb-item>商品管理</el-breadcrumb-item>
       <el-breadcrumb-item>商品列表</el-breadcrumb-item>
     </el-breadcrumb>
-<!--    卡片主体-->
+    <!--    卡片主体-->
     <el-card>
       <el-row :gutter="20">
         <el-col :span="8">
           <!--搜索框-->
-          <el-input placeholder="请输入内容" v-model="query" clearable @clear="queryClick">
-            <el-button slot="append" icon="el-icon-search" @click="queryClick"></el-button>
+          <el-input placeholder="请输入内容" v-model="query" clearable
+            @clear="queryClick">
+            <el-button slot="append" icon="el-icon-search" @click="queryClick">
+            </el-button>
           </el-input>
         </el-col>
         <el-col :span="4">
@@ -22,10 +30,14 @@
         <!--添加商品弹出框-->
       </el-row>
       <el-table :data="goodsData" style="width: 100%" stripe border>
-        <el-table-column type="index" label="#" :index="index"></el-table-column>
-        <el-table-column prop="goods_name" label="商品名称" :show-overflow-tooltip="true"></el-table-column>
-        <el-table-column prop="goods_price" label="商品价格" width="100" ></el-table-column>
-        <el-table-column prop="goods_weight" label="商品重量" width="80"></el-table-column>
+        <el-table-column type="index" label="#" :index="index">
+        </el-table-column>
+        <el-table-column prop="goods_name" label="商品名称"
+          :show-overflow-tooltip="true"></el-table-column>
+        <el-table-column prop="goods_price" label="商品价格" width="100">
+        </el-table-column>
+        <el-table-column prop="goods_weight" label="商品重量" width="80">
+        </el-table-column>
         <el-table-column prop="add_time" label="创建时间" width="150">
           <template slot-scope="scope">
             {{scope.row.add_time | dataFormat}}
@@ -33,19 +45,17 @@
         </el-table-column>
         <el-table-column label="操作" width="120">
           <template slot-scope="scope">
-            <el-button type="primary" icon="el-icon-edit" size="mini" @click="console.log(scope.row.goods_id)"></el-button>
-            <el-button type="danger" icon="el-icon-delete" size="mini" @click="deleteGoodFromId(scope.row.goods_id)"></el-button>
+            <el-button type="primary" icon="el-icon-edit" size="mini"
+              @click="console.log(scope.row.goods_id)"></el-button>
+            <el-button type="danger" icon="el-icon-delete" size="mini"
+              @click="deleteGoodFromId(scope.row.goods_id)"></el-button>
           </template>
         </el-table-column>
       </el-table>
-      <el-pagination
-        @size-change="goodsSizeChange"
-        @current-change="goodsCurrentChange"
-        :current-page="pageNum"
-        :page-sizes="[1, 2, 5, 10]"
-        :page-size="pageSize"
-        layout="total, sizes, prev, pager, next, jumper"
-        :total="total">
+      <el-pagination @size-change="goodsSizeChange"
+        @current-change="goodsCurrentChange" :current-page="pageNum"
+        :page-sizes="[1, 2, 5, 10]" :page-size="pageSize"
+        layout="total, sizes, prev, pager, next, jumper" :total="total">
       </el-pagination>
     </el-card>
   </div>
@@ -80,7 +90,7 @@ export default {
     },
     // 添加商品
     addGoodClick () {
-      console.log('addGoodClick')
+      this.$router.push('/goods/add')
     },
     // 商品可显示列
     goodsSizeChange (val) {
@@ -105,11 +115,19 @@ export default {
     },
     // 根据ID删除商品(网络请求)
     deleteGoodFromId (id) {
-      deleteGoodFromId(id).then(res => {
-        if (res.data.meta.status !== 200) return this.$message.error(res.data.meta.msg)
-        this.$message.success(res.data.meta.msg)
-        this.getGoods()
-      }).catch(err => {
+      this.$confirm('此操作将永久删除该参数, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteGoodFromId(id).then(res => {
+          if (res.data.meta.status !== 200) return this.$message.error(res.data.meta.msg)
+          this.$message.success(res.data.meta.msg)
+          this.getGoods()
+        }).catch(err => {
+          console.log(err)
+        })
+      }).catch((err) => {
         console.log(err)
       })
     }
@@ -118,5 +136,4 @@ export default {
 </script>
 
 <style scoped>
-
 </style>
